feat: embedded MongoDBManager with customizable port and timeout
Updated the MongoDBManager class to accept custom port and timeout values in its constructor, allowing for more flexibility when embedding it into other applications.

The MongoDBManager now takes two additional arguments: `port` (defaulting to 27018) and `timeout` (defaulting to 5 seconds), which can be used to customize the behavior of the embedded MongoDB instance. This change is part of an effort to make the library more versatile and easier to integrate into various use cases.