feat: Add Gwasstudio CLI Test Script and related files
Added a new script, `cli_test.sh`, which automates testing of the Gwasstudio CLI tool by performing data ingestion, querying, and exporting operations. This script provides a convenient way to test the core functionalities of Gwasstudio but, it serves also as a form of documentation, demonstrating real-world usage scenarios and providing a concrete example of how to perform common tasks with Gwasstudio .

The following changes were made:

* Added new files:
  * `scripts/README.md`
  * `scripts/cli_test.sh`
  * Updated existing files:
    + `data/search_ukb_d.txt`
    + `data/search_ukb_d_filter_by_trait_desc.txt`

The changes in this commit introduce a new feature, namely the ability to automate testing of Gwasstudio's CLI tool. This enhances the development experience and ensures that the tool works as expected