fix: switch to delayed() method for reading snp list file
Switched the way we read the snp list file from a regular function call to using the `delayed()` method in Dask. This change allows us to take advantage of parallel processing and improves performance.

This avoid embedding the large file into the graph, Dask can just pass a reference to the delayed object around.